wgc: move wgni outside api
